refactor scf methods into separate files
