+++ forte2/scf/__init__.py
@@ -1,1 +1,6 @@
-from .hf import RHF, ROHF, UHF, CUHF, GHF, convert_coeff_spatial_to_spinor
+from .rhf import RHF
+from .rohf import ROHF
+from .uhf import UHF
+from .cuhf import CUHF
+from .ghf import GHF
+from .scf_utils import convert_coeff_spatial_to_spinor
